+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## New Features
 
+- add `sos_sdk_copy_file` to copy a file without overwriting
+- add `sos_sdk_overwrite_file` to copy a file with overwriting
 - add `sos_sdk_git_clone_or_pull` to pull subprojects
 
 ## Bug Fixes